qcacld-2.0: Add TX xmit to send mgmt frames in monitor mode
Register TX xmit function for sending management frames in
monitor mode. Expand RX mgmt filter parameter as
0x0~0xFFFFFFFF

CRs-Fixed: 2642180

+++ CORE/SME/inc/sme_Api.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012-2019 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2012-2020 The Linux Foundation. All rights reserved.
  *
  * Previously licensed under the ISC license by Qualcomm Atheros, Inc.
  *
@@ -309,7 +309,7 @@ struct sme_5g_band_pref_params {
 struct sme_mnt_filter_type_req{
     u_int32_t vdev_id;
     u_int16_t request_data_len;
-    u_int8_t* request_data;
+    u_int32_t* request_data;
 };
 
 /*-------------------------------------------------------------------------
